JBPM-6824: Forms & Stunner - Improve the usability for task' actors fields #4 - Refactored Filters Fatory
