+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-forms/kie-wb-common-stunner-forms-client/src/main/java/org/kie/workbench/common/stunner/forms/client/formFilters/FormFiltersProviderFactory.java
@@ -17,14 +17,32 @@
 package org.kie.workbench.common.stunner.forms.client.formFilters;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 
+import org.kie.soup.commons.validation.PortablePreconditions;
 import org.kie.workbench.common.forms.adf.engine.shared.FormElementFilter;
 import org.kie.workbench.common.stunner.core.graph.Element;
 import org.kie.workbench.common.stunner.core.graph.content.definition.Definition;
 
-public interface StunnerFilterProviderManager {
+public class FormFiltersProviderFactory {
 
-    void init();
+    private static final Map<Class<?>, StunnerFormElementFilterProvider> providers = new HashMap<>();
 
-    Collection<FormElementFilter> getFilterForDefinition(String elementUUID, Element<? extends Definition<?>> element, Object definition);
+    public static void registerProvider(StunnerFormElementFilterProvider provider) {
+        PortablePreconditions.checkNotNull("provider", provider);
+
+        providers.put(provider.getDefinitionType(), provider);
+    }
+
+    public static Collection<FormElementFilter> getFilterForDefinition(String elementUUID, Element<? extends Definition<?>> element, Object definition) {
+        StunnerFormElementFilterProvider provider = providers.get(definition.getClass());
+
+        if (provider != null) {
+            return provider.provideFilters(elementUUID, element, definition);
+        }
+
+        return Collections.emptyList();
+    }
 }
