+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-forms/kie-wb-common-stunner-forms-client/src/test/java/org/kie/workbench/common/stunner/forms/client/widgets/container/displayer/FormDisplayerTest.java
@@ -16,12 +16,9 @@
 
 package org.kie.workbench.common.stunner.forms.client.widgets.container.displayer;
 
-import java.util.Collections;
-
 import org.jboss.errai.databinding.client.BindableProxy;
 import org.jboss.errai.databinding.client.BindableProxyFactory;
 import org.jboss.errai.databinding.client.BindableProxyProvider;
-import org.jboss.errai.ioc.client.api.ManagedInstance;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,7 +27,6 @@ import org.kie.workbench.common.forms.dynamic.service.shared.adf.DynamicFormMode
 import org.kie.workbench.common.forms.processing.engine.handling.FieldChangeHandler;
 import org.kie.workbench.common.stunner.core.graph.content.definition.Definition;
 import org.kie.workbench.common.stunner.core.graph.impl.NodeImpl;
-import org.kie.workbench.common.stunner.forms.client.formFilters.StunnerFilterProviderManager;
 import org.kie.workbench.common.stunner.forms.context.PathAwareFormContext;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
@@ -76,12 +72,6 @@ public class FormDisplayerTest {
     @Mock
     private DynamicFormModelGenerator dynamicFormModelGenerator;
 
-    @Mock
-    private ManagedInstance<StunnerFilterProviderManager> managedInstanceFilterProviderManager;
-
-    @Mock
-    private StunnerFilterProviderManager filterProviderManager;
-
     private FormDisplayer displayer;
 
     @Before
@@ -102,10 +92,8 @@ public class FormDisplayerTest {
 
         BindableProxyFactory.addBindableProxy(SecondDefinition.class,
                                               proxyProvider);
-        when(managedInstanceFilterProviderManager.get()).thenReturn(filterProviderManager);
-        when(filterProviderManager.getFilterForDefinition(any(), any(), any())).thenReturn(Collections.emptyList());
 
-        displayer = new FormDisplayer(view, formRenderer, dynamicFormModelGenerator, managedInstanceFilterProviderManager);
+        displayer = new FormDisplayer(view, formRenderer, dynamicFormModelGenerator);
 
         verify(view, times(1)).init(displayer);
     }
@@ -179,7 +167,6 @@ public class FormDisplayerTest {
         verify(formRenderer, times(initializedTimes)).isInitialized();
         verify(formRenderer, times(boundTimes)).unBind();
         verify(dynamicFormModelGenerator, times(newContextTimes)).getContextForModel(elementDefinition);
-        verify(filterProviderManager, times(newContextTimes)).getFilterForDefinition(any(), any(), any());
         verify(formRenderer, times(renderingTimes)).render(any(PathAwareFormContext.class));
         verify(formRenderer, times(renderingTimes)).addFieldChangeHandler(fieldChangeHandler);
